Resource 사용할 때 그래픽스 엔진 내부적으로 이용하던 PATH 사용 중단. -> 외부에서 경로를 아예 넣어주는것으로 교체. 그러면서 Resources 폴더 정비

+++ Floater/CubeRunner/CubeController.cpp
@@ -144,7 +144,7 @@ bool CubeController::IsOutofBoard()
 	auto pos = _gameObject->tr.GetWorldPosition();
 	auto state = _board->QueryTileState(pos.x,pos.z);
 
-	if(state == TileStateFlag::None)
+	if(state == (int)TileStateFlag::None)
 	{
 		return true;
 	}
